actionを関数として切り出し
